requirements.txt
pyyaml>=6.0
paho-mqtt>=2.1.0
